+++ src/main/java/org/itsallcode/jdbc/Transaction.java
@@ -3,8 +3,7 @@ package org.itsallcode.jdbc;
 import java.sql.Connection;
 import java.util.function.Consumer;
 
-import org.itsallcode.jdbc.batch.BatchInsertBuilder;
-import org.itsallcode.jdbc.batch.RowBatchInsertBuilder;
+import org.itsallcode.jdbc.batch.*;
 import org.itsallcode.jdbc.resultset.RowMapper;
 import org.itsallcode.jdbc.resultset.SimpleResultSet;
 import org.itsallcode.jdbc.resultset.generic.Row;
@@ -75,9 +74,15 @@ public final class Transaction implements DbOperations {
     }
 
     @Override
-    public void executeStatement(final String sql, final PreparedStatementSetter preparedStatementSetter) {
+    public int executeUpdate(final String sql) {
         checkOperationAllowed();
-        connection.executeStatement(sql, preparedStatementSetter);
+        return connection.executeUpdate(sql);
+    }
+
+    @Override
+    public int executeUpdate(final String sql, final PreparedStatementSetter preparedStatementSetter) {
+        checkOperationAllowed();
+        return connection.executeUpdate(sql, preparedStatementSetter);
     }
 
     @Override
@@ -97,6 +102,12 @@ public final class Transaction implements DbOperations {
             final RowMapper<T> rowMapper) {
         checkOperationAllowed();
         return connection.query(sql, preparedStatementSetter, rowMapper);
+    }
+
+    @Override
+    public StatementBatchBuilder batch() {
+        checkOperationAllowed();
+        return connection.batch();
     }
 
     @Override